+++ src/qbssettings.ts
@@ -50,6 +50,7 @@ const DEFAULT_SHOW_DISABLED_PROJECT_ITEMS = true;
 const DEFAULT_CLEAR_OUTPUT_BEFORE_OPERATION = false;
 const DEFAULT_SAVE_BEFORE_BUILD = true;
 const DEFAULT_AUTO_RESOLVE = true;
+const DEFAULT_BUILD_BEFORE_RUN = true;
 const DEFAULT_CONFIGURATIONS_FILE_PATH = `${SOURCE_DIR_PATTERN}/.vscode/qbs-configurations.json`;
 
 export enum QbsSettingsEvent {
@@ -213,6 +214,10 @@ export class QbsSettings implements vscode.Disposable {
 
     autoResolve(): boolean {
         return this._settings.get<boolean>('autoResolve', DEFAULT_AUTO_RESOLVE);
+    }
+
+    buildBeforeRun(): boolean {
+        return this._settings.get<boolean>('buildBeforeRun', DEFAULT_BUILD_BEFORE_RUN);
     }
 
     /**